Handle not found for update, delete EventSubmissionController.cs

+++ backend/Controllers/EventSubmissionController.cs
@@ -50,6 +50,12 @@ namespace backend.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(string id, EventSubmission eventSubmission)
         {
+            var exist = await _eventSubmissionRepository.Get(id);
+            if (exist == null)
+            {
+                return NotFound();
+            }
+            
             var result = await _eventSubmissionRepository.Update(id, eventSubmission);
             return new JsonResult(result);
         }
@@ -57,6 +63,12 @@ namespace backend.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            var exist = await _eventSubmissionRepository.Get(id);
+            if (exist == null)
+            {
+                return NotFound();
+            }
+            
             var result = await _eventSubmissionRepository.Delete(id);
             return new JsonResult(result);
         }
